refactor: consolidate text utilities in inventory/page-layout domains
- Extract firstNonBlank, firstString to core/utils/text.ts (LOTE 1)
  * Eliminates 2 duplicate implementations in assemble/layout-shared
  * Improves code reuse across inventory/page-layout domains

- Extract trimLeadingSlash helper to core/utils/text.ts (LOTE 2)
  * Consolidates 7 instances of startsWith('/') ? slice(1) : value pattern
  * Applied across: admin-urls, layout-export, page-fetch, inventory-pages
  * Improves readability and maintainability

Net result: -23 lines of code, -2 duplicated functions, 0 regressions
Tests: 514 passing, no new lint warnings
